updated cloud moves at slow step, updated form sprite renderer to take in float values and draw Forms poses to properly interpolate between Cells

+++ formglfw/FormGlfw.h
@@ -14,11 +14,11 @@ typedef struct AnimOrder {
 
 void updateLoop();
 void drawWorld(World *w, int tMat, int sMat, int rMat, int color, GLuint squa, int sTrans, int sScale, int sRot);
-void drawFormSprite(Form *f, float *sMatrix, float xSize, float ySize, int xp, int yp, GLuint sScale, GLuint sTrans, GLuint sRot);
+void drawFormSprite(Form *f, float *sMatrix, float xSize, float ySize, float xp, float yp, GLuint sScale, GLuint sTrans, GLuint sRot);
 void drawGrid(float *mat, int tMat, int sMat,  int rMat, int color, GLuint vLi);
 //void keyCallback(GLFWwindow *window, int key, int scancode, int action, int mods);
 AnimOrder *makeAnimOrder(int order);
-void addFormToAnim(AnimOrder *ao, Form *f, int x, int y);
+void addFormToAnim(AnimOrder *ao, Form *f, float x, float y);
 void drawAnimOrder(AnimOrder *ao, float *sMatrix, float xSize, float ySize,  GLuint sScale, GLuint sTrans, GLuint sRot);
 void freeAnimOrder(AnimOrder *ao);
 void setCenter(float cp[2]);
